+++ src/components/ui/modals/DialogSystem.jsx
@@ -6,9 +6,8 @@ import { Box, Typography, useTheme } from '@mui/material';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import { Button } from '../../components/ui/buttons';
-
 import Dialog from './Dialog';
+import { Button } from '../buttons';
 
 /**
  * DialogSystem component that provides a standardized way to display different types of dialogs
